Updated the README to add ansible_connection
The ansible_connection: local was missing from this file. As such when running the job an unreachable error would be generated. This line was already in the original file and the instructions just needed a small modification to ensure people left the ansible_connection line present. As a result now copying the entire content for ci-cd-tooling.yml works as expected with the change made.